+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/package.json
@@ -12,10 +12,10 @@
         "clean": "rm -rf node_modules .dist",
         "cypress-open": "cypress open --project $PWD/e2e",
         "cypress-run": "cypress run --project $PWD/e2e",
-        "dev-api": "PORT=8089 node ./server/api-mock.js",
+        "dev-api": "PORT=8089 node ./server/api-mock.cjs",
         "lint": "eslint ./src --ext ts,tsx",
         "snd": "cross-env NODE_ENV=development node ./server.cjs",
-        "start-api-mock": "PORT=8089 node ./server/api-mock.js",
+        "start-api-mock": "PORT=8089 node ./server/api-mock.cjs",
         "start-e2e-server": "node ./server-e2e.cjs",
         "start-express": "node ./server.cjs",
         "test-watch": "react-scripts test --watchAll=true --env=jsdom",
