+++ packages/sif-server-utils/index.cjs
@@ -1,15 +1,14 @@
-const express = require('express');
+const { createProxyMiddleware } = require('http-proxy-middleware');
+const { initTokenX, exchangeToken } = require('./tokenx.cjs');
 const { v4: uuidv4 } = require('uuid');
-const jose = require('jose');
-const mustacheExpress = require('mustache-express');
 const compression = require('compression');
 const cookieParser = require('cookie-parser');
-const RateLimit = require('express-rate-limit');
-
+const express = require('express');
 const getDecorator = require('./decorator.cjs');
-const { initTokenX, exchangeToken } = require('./tokenx.cjs');
-
-const isDev = process.env.NODE_ENV === 'development';
+const jose = require('jose');
+const mustacheExpress = require('mustache-express');
+const path = require('path');
+const RateLimit = require('express-rate-limit');
 
 const isExpiredOrNotAuthorized = (token) => {
     if (token) {
@@ -25,10 +24,10 @@ const isExpiredOrNotAuthorized = (token) => {
     return true;
 };
 
-const getRouterConfig = async (req, audienceInnsyn) => {
+const getRouterConfig = async (NAIS_CLIENT_ID, req, audienceInnsyn) => {
     req.headers['X-Correlation-ID'] = uuidv4();
-    if (process.env.NAIS_CLIENT_ID !== undefined) {
-        req.headers['X-K9-Brukerdialog'] = process.env.NAIS_CLIENT_ID;
+    if (NAIS_CLIENT_ID !== undefined) {
+        req.headers['X-K9-Brukerdialog'] = NAIS_CLIENT_ID;
     }
     if (req.headers['authorization'] !== undefined) {
         const token = req.headers['authorization'].replace('Bearer ', '');
@@ -54,15 +53,15 @@ const renderApp = (decoratorFragments, server) =>
         });
     });
 
-const setupTokenX = async () => {
+const setupTokenX = async (isDev) => {
     if (isDev) {
         return Promise.resolve();
     }
     return Promise.all([initTokenX()]);
 };
 
-const getDecoratorAndServer = async (AppSettings, dirName) => {
-    const server = getSifServer(dirName);
+const getDecoratorAndServer = async (AppSettings, dirName, isDev) => {
+    const server = getSifServer(dirName, isDev);
     return getDecorator(AppSettings)
         .then(
             (decoratorFragments) => renderApp(decoratorFragments, server),
@@ -73,7 +72,7 @@ const getDecoratorAndServer = async (AppSettings, dirName) => {
         )
         .then(
             async (html) => {
-                await setupTokenX();
+                await setupTokenX(isDev);
                 return {
                     html,
                     server,
@@ -83,7 +82,7 @@ const getDecoratorAndServer = async (AppSettings, dirName) => {
         );
 };
 
-const getSifServer = (dirname) => {
+const getSifServer = (dirname, isDev) => {
     const server = express();
     server.use((_req, res, next) => {
         res.removeHeader('X-Powered-By');
@@ -104,22 +103,65 @@ const getSifServer = (dirname) => {
     }
     server.set('view engine', 'mustache');
 
-    if (isDev) {
-        server.use(
-            '/api',
-            RateLimit({
-                windowMs: 15 * 60 * 1000, // 15 minutes
-                max: 500, // max 100 requests per windowMs
-                standardHeaders: 'draft-7',
-                legacyHeaders: false,
-            }),
-        );
-    }
-
     return server;
 };
 
+const createApiUrlProxyMiddleware = (target, replacement, NAIS_CLIENT_ID) =>
+    createProxyMiddleware({
+        target,
+        changeOrigin: true,
+        pathRewrite: (path) => {
+            return path.replace(replacement, '');
+        },
+        router: async (req) => getRouterConfig(NAIS_CLIENT_ID, req, false),
+        secure: true,
+        xfwd: true,
+        logLevel: 'info',
+    });
+
+const setupViteDevServer = async (server, dirname, html) => {
+    const fs = require('fs');
+    fs.writeFileSync(path.resolve(dirname, 'index-decorated.html'), html);
+    const vedleggMockStore = './dist/vedlegg';
+
+    if (!fs.existsSync(vedleggMockStore)) {
+        fs.mkdirSync(vedleggMockStore);
+    }
+
+    const vite = await require('vite').createServer({
+        root: dirname,
+        server: {
+            middlewareMode: true,
+            port: 8080,
+            open: './index-decorated.html',
+        },
+    });
+
+    server.use(
+        '/api',
+        RateLimit({
+            windowMs: 15 * 60 * 1000, // 15 minutes
+            max: 500, // max 100 requests per windowMs
+            standardHeaders: 'draft-7',
+            legacyHeaders: false,
+        }),
+    );
+
+    server.get(/^\/(?!.*dist).*$/, (req, _res, next) => {
+        const fullPath = path.resolve(dirname, decodeURIComponent(req.path.substring(1)));
+        const fileExists = fs.existsSync(fullPath);
+
+        if ((!fileExists && !req.url.startsWith('/@')) || req.url === '/') {
+            req.url = '/index-decorated.html';
+        }
+        next();
+    });
+
+    server.use(vite.middlewares);
+};
 module.exports = {
+    setupViteDevServer,
+    createApiUrlProxyMiddleware,
     getDecoratorAndServer,
     getRouterConfig,
 };
