+++ packages/sif-server-utils/package.json
@@ -8,6 +8,7 @@
         "cookie-parser": "1.4.6",
         "express": "4.18.2",
         "express-rate-limit": "7.1.3",
+        "http-proxy-middleware": "2.0.6",
         "jose": "5.0.1",
         "mustache-express": "1.3.2",
         "uuid": "9.0.1"
